+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Development]
+Nothing yet!
+
+## [v0.5.0] - 2015-10-28
 ### Added
 - getSecrets() method for more efficient secret retrieval batching
 
@@ -58,7 +61,8 @@ secret.
 ### Added
 - Initial release
 
-[Development]: https://github.com/TechnologyAdvice/Cryptex/compare/0.4.1...HEAD
+[Development]: https://github.com/TechnologyAdvice/Cryptex/compare/0.5.0...HEAD
+[v0.5.0]: https://github.com/TechnologyAdvice/Cryptex/compare/0.4.1...0.5.0
 [v0.4.1]: https://github.com/TechnologyAdvice/Cryptex/compare/0.4.0...0.4.1
 [v0.4.0]: https://github.com/TechnologyAdvice/Cryptex/compare/0.3.1...0.4.0
 [v0.3.1]: https://github.com/TechnologyAdvice/Cryptex/compare/0.3.0...0.3.1